Update evil-portal.ino - blank the screen to avoid burn-ins
With the Cardputer demo software, I noticed a slight burn-in on the screen, when the same image is shown over a rather short period of time.

In order to avoid burn-in when using Evil Portal over a few hours, I added three lines to clear the screen 5 seconds after showing each new captured credencial. This way I hope to extend the life of the screen.